requirements-netlify.txt
# Requirements for the Netlify portion of the deploy
google-cloud-storage